+++ CHANGELOG.md
@@ -1,7 +1,15 @@
 # BedWar Mod - Change Log
 
-## Version 0.1
+# Version 0.1.2
+- Various bug fixes
+- Formatting improvements for stats features
 
+# Version 0.1.1
+- Made all the chat events work with the repo
+- Change some wording of features
+- Tab complete in BedWars games now bases the players off of when the game started
+
+## Version 0.1
 - Setup GitHub repository
 - Setup project - Auto update, fetching from repo, config menu
 - Resource overlay
@@ -13,8 +21,7 @@
 - Party commands shortening and tab completing
 - Ping on chat mention
 - Copy chat messages
-
-### `/bws <player>` command
+- `/bws <player>` command
 - See stats of players
 - Can switch between game modes
 - Auto send on player joining your party